Implement ISessionLoggerScopeProvider to supply session context to loggers
The provider owns a field for the current session ID that can be updated via the SetCurrentSessionId() method. Callers can then use BeginSessionScope() with their desired ILogger to begin scope with the session ID attached.

The Session.ID type is changed from GUID to a string generated by the Nanoid package.